requirements.txt
cssmin
# #328 Flask-Images break with itsdangerous>=2.0
# In turn flash >= 2.0 break with itsdangerous < 2.0
itsdangerous<2.0
flask<2.0
Werkzeug<2.0
flask-assets
flask-login
sqlalchemy<1.4
flask_sqlalchemy
sqlalchemy-utils
passlib
libsass
flask-wtf
WTForms-Alchemy
# See #98
git+https://github.com/maxcountryman/flask-uploads.git@f66d7dc93e684fa0a3a4350a38e41ae00483a796#egg=flask-uploads
flask-migrate
flask-marshmallow
marshmallow-sqlalchemy
Flask-Images
markdown
# Grab pysimplesoap from Github to get list serialization fixes
git+https://github.com/pysimplesoap/pysimplesoap.git@ad03a21cafab982eed321553c4bfcda1755182eb#egg=pysimplesoap
pymysql
dkimpy
email_validator
openpyxl
babel
flask-statistics
anonymizeip
python-dateutil
